+++ src/main/java/edu/ntnu/idatt2106/boco/service/UserService.java
@@ -4,6 +4,7 @@ import edu.ntnu.idatt2106.boco.models.Image;
 import edu.ntnu.idatt2106.boco.models.User;
 import edu.ntnu.idatt2106.boco.payload.request.LoginRequest;
 import edu.ntnu.idatt2106.boco.payload.request.RegisterUserRequest;
+import edu.ntnu.idatt2106.boco.payload.request.UpdateUserAdminRequest;
 import edu.ntnu.idatt2106.boco.payload.request.UpdateUserRequest;
 import edu.ntnu.idatt2106.boco.payload.response.UserResponse;
 import edu.ntnu.idatt2106.boco.repository.ImageRepository;
@@ -12,7 +13,10 @@ import edu.ntnu.idatt2106.boco.util.Mapper;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
-import org.springframework.web.multipart.MultipartFile;
+
+import java.util.List;
+import java.util.Locale;
+import java.util.Objects;
 import java.util.Optional;
 
 @Service
@@ -117,7 +121,10 @@ public class UserService
         if (request.getImageId() != null)
         {
             Image prevImage = user.getImage();
-            imageRepository.delete(prevImage);
+            if (prevImage != null && !Objects.equals(request.getImageId(), prevImage.getImageId()))
+            {
+                imageRepository.delete(prevImage);
+            }
 
             Optional<Image> optionalImage = imageRepository.findById(request.getImageId());
             if (optionalImage.isPresent()) user.setImage(optionalImage.get());
@@ -125,6 +132,24 @@ public class UserService
 
         user = userRepository.save(user);
         return Mapper.ToUserResponse(user);
+    }
+
+    public UserResponse updateUserRoleAdmin(UpdateUserAdminRequest request)
+    {
+        Optional<User> optionalUser = userRepository.findById(request.getUserId());
+        if (optionalUser.isEmpty()) return null;
+        User user = optionalUser.get();
+
+        if (request.getRole() != null) user.setRole(request.getRole().toUpperCase(Locale.ROOT));
+
+        user = userRepository.save(user);
+        return Mapper.ToUserResponse(user);
+    }
+
+    public List<UserResponse> getAllUsers()
+    {
+        List<User> users = userRepository.findAll();
+        return Mapper.ToUserResponses(users);
     }
 }
